+++ precompiles/vesting/tx_test.go
@@ -8,9 +8,7 @@ import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/ethereum/go-ethereum/core/vm"
 	cmn "github.com/evmos/evmos/v14/precompiles/common"
-	"github.com/evmos/evmos/v14/precompiles/testutil"
 	"github.com/evmos/evmos/v14/precompiles/vesting"
 	evmosutil "github.com/evmos/evmos/v14/testutil"
 	evmosutiltx "github.com/evmos/evmos/v14/testutil/tx"
@@ -98,10 +96,7 @@ func (s *PrecompileTestSuite) TestCreateClawbackVestingAccount() {
 		s.Run(tc.name, func() {
 			s.SetupTest()
 
-			var contract *vm.Contract
-			contract, s.ctx = testutil.NewPrecompileContract(s.T(), s.ctx, s.address, s.precompile, tc.gas)
-
-			bz, err := s.precompile.CreateClawbackVestingAccount(s.ctx, s.address, contract, s.stateDB, &method, tc.malleate())
+			bz, err := s.precompile.CreateClawbackVestingAccount(s.ctx, s.address, s.stateDB, &method, tc.malleate())
 
 			if tc.expError {
 				s.Require().ErrorContains(err, tc.errContains)
@@ -185,10 +180,7 @@ func (s *PrecompileTestSuite) TestFundVestingAccount() {
 		s.Run(tc.name, func() {
 			s.SetupTest()
 
-			var contract *vm.Contract
-			contract, s.ctx = testutil.NewPrecompileContract(s.T(), s.ctx, s.address, s.precompile, tc.gas)
-
-			bz, err := s.precompile.FundVestingAccount(s.ctx, s.address, contract, s.stateDB, &method, tc.malleate())
+			bz, err := s.precompile.FundVestingAccount(s.ctx, s.address, s.stateDB, &method, tc.malleate())
 
 			if tc.expError {
 				s.Require().ErrorContains(err, tc.errContains)
@@ -263,10 +255,7 @@ func (s *PrecompileTestSuite) TestClawback() {
 		s.Run(tc.name, func() {
 			s.SetupTest()
 
-			var contract *vm.Contract
-			contract, s.ctx = testutil.NewPrecompileContract(s.T(), s.ctx, s.address, s.precompile, tc.gas)
-
-			bz, err := s.precompile.Clawback(s.ctx, s.address, contract, s.stateDB, &method, tc.malleate())
+			bz, err := s.precompile.Clawback(s.ctx, s.address, s.stateDB, &method, tc.malleate())
 
 			if tc.expError {
 				s.Require().ErrorContains(err, tc.errContains)
@@ -349,10 +338,7 @@ func (s *PrecompileTestSuite) TestUpdateVestingFunder() {
 		s.Run(tc.name, func() {
 			s.SetupTest()
 
-			var contract *vm.Contract
-			contract, s.ctx = testutil.NewPrecompileContract(s.T(), s.ctx, s.address, s.precompile, tc.gas)
-
-			bz, err := s.precompile.UpdateVestingFunder(s.ctx, s.address, contract, s.stateDB, &method, tc.malleate())
+			bz, err := s.precompile.UpdateVestingFunder(s.ctx, s.address, s.stateDB, &method, tc.malleate())
 
 			if tc.expError {
 				s.Require().ErrorContains(err, tc.errContains)