+++ precompiles/vesting/vesting.go
@@ -98,18 +98,18 @@ func (p Precompile) Run(evm *vm.EVM, contract *vm.Contract, readOnly bool) (bz [
 	switch method.Name {
 	// Vesting transactions
 	case CreateClawbackVestingAccountMethod:
-		bz, err = p.CreateClawbackVestingAccount(ctx, evm.Origin, contract, stateDB, method, args)
+		bz, err = p.CreateClawbackVestingAccount(ctx, evm.Origin, stateDB, method, args)
 	case FundVestingAccountMethod:
-		bz, err = p.FundVestingAccount(ctx, evm.Origin, contract, stateDB, method, args)
+		bz, err = p.FundVestingAccount(ctx, evm.Origin, stateDB, method, args)
 	case ClawbackMethod:
-		bz, err = p.Clawback(ctx, evm.Origin, contract, stateDB, method, args)
+		bz, err = p.Clawback(ctx, evm.Origin, stateDB, method, args)
 	case UpdateVestingFunderMethod:
-		bz, err = p.UpdateVestingFunder(ctx, evm.Origin, contract, stateDB, method, args)
+		bz, err = p.UpdateVestingFunder(ctx, evm.Origin, stateDB, method, args)
 	case ConvertVestingAccountMethod:
 		bz, err = p.ConvertVestingAccount(ctx, stateDB, method, args)
 	// Vesting queries
 	case BalancesMethod:
-		bz, err = p.Balances(ctx, contract, method, args)
+		bz, err = p.Balances(ctx, method, args)
 	}
 
 	if err != nil {