+++ x/revenue/v1/migrations/v2/types/genesis.pb.go
@@ -11,8 +11,8 @@ import (
 
 	github_com_cosmos_cosmos_sdk_types "github.com/cosmos/cosmos-sdk/types"
 	_ "github.com/cosmos/gogoproto/gogoproto"
-	"github.com/evmos/evmos/v14/x/revenue/v1/types"
 	proto "github.com/cosmos/gogoproto/proto"
+	"github.com/evmos/evmos/v14/x/revenue/v1/types"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.